% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baytsDD.R
\name{baytsDD}
\alias{baytsDD}
\title{baytsDD}
\usage{
baytsDD(tsL = list(NULL, ...), msdL = list(), distNFL = list(), start,
  end = NULL, formula = response ~ trend + harmon, order = 1, chi = 0.9,
  PNFmin = 0.5, bwf = c(0.1, 0.9), residuals = FALSE)
}
\arguments{
\item{tsL}{list of object(s) of class \code{\link{ts}}.}

\item{msdL}{list of msdl object(s) describing the modulation of the sd of F and NF sd(F),sd(NF),mean(NF) (e.g. 2,2,-4)}

\item{start}{Start date of monitoring period. Default=NULL (start of input time series). Period before is used for deseasonalisation}

\item{formula}{formula for the regression model. The default is response ~ trend + harmon, i.e., a linear trend and a harmonic season component. Other specifications are possible using all terms set up by bfastpp, i.e., season (seasonal pattern with dummy variables), lag (autoregressive terms), slag (seasonal autoregressive terms), or xreg (further covariates). See bfastpp for details.}

\item{order}{numeric. Order of the harmonic term, defaulting to 3.}

\item{residuals}{TRUE = output are time series of deseasonalised "residuals"; FALSE = time series of deseasonalised values}

\item{pdfL}{list of "pdf" object(s) describing F and NF distributions (see \code{\link{calcPNF}}).}

\item{distL}{list of "distNF" object(s) describing the mean and sd of the NF distribution in case no data driven way to derive the NF distribution is wanted; default=NULL}
}
\value{
List of 7 output paramter.
(1) bayts: "bayts" time series data frame
(2) flag: time at which unconfirmed change got flagged;
(3) change.flagged: time at which confirmed change got flagged;
(4) change.confirmed: time at which change is confirmed;
(5) oldflag: time of earlier flagged but not confirmed changes;
(6) vchange: vector of time steps from time at which change got flagged until confirmation;
(7) vflag: vector of time steps at which unconfirmed change is flaged
}
\description{
(i) deseasonalising time series using harmonic model fitting , (ii) deriving F and NF pdfs in data driven way and (iii) run bayts to detect forest chang
}
\author{
Johannes Reiche (Wageningen University)
}

