% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baytsDDSpatial.R
\name{baytsDDSpatial}
\alias{baytsDDSpatial}
\title{Function to run baytsDD on (mulitple) raster bricks}
\usage{
baytsDDSpatial(bL = list(), datesL = list(), pdfsdL = list(),
  distNFL = list(), modL = list(), formulaL = list(), orderL = list(),
  mask = NULL, start_history = NULL, end_history = NULL, start,
  end = NULL, chi = 0.9, PNFmin = 0.5, bwf = c(0.1, 0.9),
  mc.cores = 1, out_file = NULL)
}
\arguments{
\item{bL}{list of raster bricks. Raster bricks need to have the same extent and spatial resolution.}

\item{datesL}{list of time vector of the format: "2014-10-07".}

\item{pdfsdL}{list of pdfsd object(s) describing the modulation of the sd of F and NF sd(F),mean(NF),sd(NF) (e.g. pdfsd = c(2,-4,2))}

\item{modL}{list of modL - modulation of the time series observations. default=NULL}

\item{formulaL}{list of formula for the regression model. The default is response ~ trend + harmon, i.e., a linear trend and a harmonic season component. Other specifications are possible using all terms set up by bfastpp, i.e., season (seasonal pattern with dummy variables), lag (autoregressive terms), slag (seasonal autoregressive terms), or xreg (further covariates). See bfastpp for details.}

\item{orderL}{list of numeric. Order of the harmonic term, defaulting to 3.}

\item{mask}{(raster) mask at which method is applied; default = NULL (method is applied to all pixel)}

\item{start_history}{Start date of history period used to model the seasonality and derive F and NF PDFs. Default=NULL (start of input time series)}

\item{end_history}{End date of history period used to model the seasonality and derive F and NF PDFs. Default=NULL (Start of the monitoring period is used)}

\item{start}{start date of monitoring period. Default=NULL (start of input time series).}

\item{end}{end date of monitoring period. Default=NULL (end of input time series)}

\item{chi}{threshold of Pchange at which the change is confirmed; Default=0.5}

\item{PNFmin}{threshold of pNF above which the first observation is flagged; Default=0.5}

\item{bwf}{block weighting function to truncate the NF probability; Default=c(0.1,0.9); (c(0,1) = no truncation)}

\item{mc.cores}{numeric. number of cores to be used for the job. See \code{\link{mc.calc}} for more details (default = 1)}

\item{distL}{list of "distNF" object(s) describing the mean and sd of the NF distribution in case no data driven way to derive the NF distribution is wanted; default=NULL}

\item{outfile}{output file}
}
\value{
A rasterBrick with 5 layers:
(1) flag: time at which unconfirmed change got flagged;
(2) change.flagged: time at which confirmed change got flagged;
(3) change.confirmed: time at which change is confirmed;
(4) Pflag: Probabilty of change for unconfirmed flagged changes;
(5) Pchange.confirmed: Probabilty of change for confirmed changes.
}
\description{
Implements baytsDD function on (multiple) time series rasterBrick object(s).
}
\examples{
#TBD
}
\author{
Johannes Reiche (Wageningen University)
}
\references{
\href{http://www.mdpi.com/2072-4292/7/5/4973}{Reiche et al. (2015): A Bayesian Approach to Combine Landsat and ALOS PALSAR Time Series for Near Real-Time Deforestation Detection. Remote Sensing. 7(5), 4973-4996; doi:10.3390/rs70504973}
}

