% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baytsSpatial.R
\name{baytsSpatial}
\alias{baytsSpatial}
\title{Function to run bayts on mulitple raster bricks}
\usage{
baytsSpatial(bL = list(NULL, ...), datesL = list(NULL, ...),
  pdfL = list(NULL, ...), bwf = c(0.1, 0.9), chi = 0.9, PNFmin = 0.5,
  start = NULL, end = NULL, pptype = "irregular", out_file = NULL,
  mc.cores = 1)
}
\arguments{
\item{bL}{list of raster bricks. Raster bricks need to have the same extent and spatial resolution.}

\item{datesL}{list of time vector of the format: "2014-10-07".}

\item{pdfL}{list of "pdf" object(s) describing F and NF distributions (see \code{\link{calcPNF}}).
"pdf" object: pdf[1] = pdf type F, pdf[2] = pdf type NF, pdf[3] and pdf[4] = pdf parameter describing F, pdf[5] and pdf[6] = pdf parameter describing NF. pdf types supported: Gaussian or Weibull.}

\item{bwf}{block weighting function to truncate the NF probability; Default=c(0.1,0.9); (c(0,1) = no truncation)}

\item{chi}{threshold of Pchange at which the change is confirmed; Default=0.5}

\item{PNFmin}{threshold of pNF above which the first observation is flagged; Default=0.5}

\item{start}{start date of monitoring period. Default=NULL (start of input time series).}

\item{end}{end date of monitoring period. Default=NULL (end of input time series)}

\item{pptype}{character. Type of preprocessing to be applied to individual time series vectors. The two options are 'irregular' and '16-days'. See \code{\link{bfastts}} for more details.}

\item{mc.cores}{numeric. number of cores to be used for the job. See \code{\link{mc.calc}} for more details (default = 1)}

\item{outfile}{output file}
}
\value{
A rasterBrick with 5 layers:
(1) flag: time at which unconfirmed change got flagged;
(2) change.flagged: time at which confirmed change got flagged;
(3) change.confirmed: time at which change is confirmed;
(4) Pflag: Probabilty of change for unconfirmed flagged changes;
(5) Pchange.confirmed: Probabilty of change for confirmed changes.
}
\description{
Implements bayts function on multiple time series rasterBrick object(s).
Time information is provided as an extra object and the time series can be regular or irregular. Information describing F and NF distributions is provided for each time series. See (\code{\link{bayts}}) for more details.
}
\examples{
#TBD
}
\author{
Johannes Reiche (Wageningen University)
}
\references{
\href{http://www.mdpi.com/2072-4292/7/5/4973}{Reiche et al. (2015): A Bayesian Approach to Combine Landsat and ALOS PALSAR Time Series for Near Real-Time Deforestation Detection. Remote Sensing. 7(5), 4973-4996; doi:10.3390/rs70504973}
}

