% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcPosterior.R
\name{calcPosterior}
\alias{calcPosterior}
\title{Calculating posterior (Bayes theorem)}
\usage{
calcPosterior(prior, likelihood)
}
\arguments{
\item{prior}{single value, vector or time series object of class "ts"}

\item{likelihood}{single value, vector or time series object of class "ts"}
}
\value{
single value, vector or time series object of class "ts"
}
\description{
Calculating posterior according to Bayes theorem.
}
\author{
Johannes Reiche (Wageningen University)
}
\references{
\href{http://www.mdpi.com/2072-4292/7/5/4973}{Reiche et al. (2015): A Bayesian Approach to Combine Landsat and ALOS PALSAR Time Series for Near Real-Time Deforestation Detection. Remote Sensing. 7(5), 4973-4996; doi:10.3390/rs70504973}
}

