% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createBayts.R
\name{createBayts}
\alias{createBayts}
\title{Creates "bayts" time series data frame for multiple input time series}
\usage{
createBayts(tsL = list(NULL, ...), pdfL = list(NULL, ...), bwf = c(0.1,
  0.9))
}
\arguments{
\item{tsL}{list of object(s) of class \code{\link{ts}}}

\item{pdfL}{list of "pdf" object(s) describing F and NF distributions (see \code{\link{calcPNF}}).
"pdf" object: pdf[1] = pdf type F, pdf[2] = pdf type NF, pdf[3] and pdf[4] = pdf parameter describing F, pdf[5] and pdf[6] = pdf parameter describing NF. pdf types supported: Gaussian or Weibull.}

\item{bwf}{block weighting function to truncate the NF probability (default = c(0,1) = no truncation)}
}
\value{
"bayts" time seris data frame including: time series observation (ts1, ts2 ...), conditional non-forest probabilties (PNF), empty Flag (Flag) and empty Change probability (PChange).
}
\description{
Creates "bayts" time series data frame from multiple input time series and calculates time series of normalised conditional non-forest probabilities (PNF).
}
\details{
Short method description: PNF is estimated for each individual time series observation, using the corresponding sensor specific probability density functions (pdfs) for forest (F) and non-forest (NF).
In case of multiple observation at the same date, PNF is interatively updated using Bayesian probability updating (\code{\link{calcPosterior}}).
A detailed description is provided in  \href{http://www.mdpi.com/2072-4292/7/5/4973}{Reiche et al. 2015} (Chapter 2.1.2 and 2.1.3).
}
\examples{
#TBD
}
\author{
Johannes Reiche (Wageningen University)
}
\references{
\href{http://www.mdpi.com/2072-4292/7/5/4973}{Reiche et al. (2015): A Bayesian Approach to Combine Landsat and ALOS PALSAR Time Series for Near Real-Time Deforestation Detection. Remote Sensing. 7(5), 4973-4996; doi:10.3390/rs70504973}
}

