% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deseasonalizeRaster.R
\name{deseasonalizeRaster}
\alias{deseasonalizeRaster}
\title{Deseasonlize a raster}
\usage{
deseasonalizeRaster(b, percentile = 0.95, min = NULL, max = NULL)
}
\arguments{
\item{b}{raster or raster brick}

\item{percentile}{percentile of the distribution used to deseasonalize the raster. Values between 0 - 1.}

\item{min}{minimum value allowed. Values below the minimum are set to NA.}

\item{max}{maximum value allowed. Values above the maximum are set to NA.}
}
\value{
deseasonlized raster layer or raster brick
}
\description{
Deseasonlize a single raster layer or each individual raster layer in a brick. "Deseasonlized pixel value" = "original pixel value" - "percentile of the distribution of the raster"
}
\examples{
data(s1vv_lndvi_raster)

plot(s1vv,3)

s1vvd <- deseasonalizeRaster(s1vv,p=0.95)

plot(s1vvd,3)
}
\author{
Johannes Reiche (Wageningen University)
}

