% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deseasonalizeTSpdf.R
\name{deseasonalizeTSpdf}
\alias{deseasonalizeTSpdf}
\title{Deseasonlize a time series using harmonic model fitting and deriving pdfs for forest and non-forest}
\usage{
deseasonalizeTSpdf(tsL = list(NULL, ...), distNFL = list(), msdL = list(),
  start, formula = response ~ trend + harmon, order = 1,
  residuals = FALSE)
}
\arguments{
\item{tsL}{list of object(s) of class \code{\link{ts}}.}

\item{msdL}{list of msdl object(s) describing the modulation of the sd of F and NF sd(F),sd(NF),mean(NF) (e.g. 2,2,-4)}

\item{start}{Start date of monitoring period. Default=NULL (start of input time series).}

\item{formula}{formula for the regression model. The default is response ~ trend + harmon, i.e., a linear trend and a harmonic season component. Other specifications are possible using all terms set up by bfastpp, i.e., season (seasonal pattern with dummy variables), lag (autoregressive terms), slag (seasonal autoregressive terms), or xreg (further covariates). See bfastpp for details.}

\item{order}{numeric. Order of the harmonic term, defaulting to 3.}

\item{residuals}{TRUE = output are time series of deseasonalised "residuals"; FALSE = time series of deseasonalised values}

\item{pdfL}{list of "pdf" object(s) describing F and NF distributions (see \code{\link{calcPNF}}).}

\item{distL}{list of "distNF" object(s) describing the mean and sd of the NF distribution in case no data driven way to derive the NF distribution is wanted; default=NULL}
}
\value{
deseasonlized time series (residuals or real values) and Gaussian pdfs of defining F and NF distribution
}
\description{
Method as described in Reiche et al., 2018 (Remote Sensing)
}
\author{
Johannes Reiche (Wageningen University)
}

