% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectBayts.R
\name{detectBayts}
\alias{detectBayts}
\title{Calculate change probability and detect changes}
\usage{
detectBayts(bayts, chi = 0.5, start = NULL, end = NULL)
}
\arguments{
\item{bayts}{"bayts" time series data frame created with \code{\link{createBayts}}}

\item{chi}{Threshold of Pchange at which the change is confirmed; Default=0.5}

\item{start}{Start date of monitoring period. Default=NULL (start of "bayts" time series data frame).}

\item{end}{End date of monitoring period. Default=NULL (end of "bayts" time series data frame)}
}
\value{
Updated "bayts" time series data frame with changes if detected.

Flag: "0" = no change flagged;
"Flag" = change flagged (iterative Bayesian updating is ongoing); "oldFlag" = old flagged change that was rejected; "Change" = confirmed change and observations that were initially flagged as change
}
\description{
Iterative Bayesian updating of the conditional probability of change (PChange) based on PNF and detection of change
}
\details{
Short method description: Observations at time t (current observation) are flagged to be potentially changed in the case that the conditional NF probability (PNF) is larger than 0.5.
For a flagged observation, the conditional probability of change (PChange) is computed by iterative Bayesian updating (\code{\link{calcPosterior}}), using the previous observation (t − 1),
the current observation (t), as well as i upcoming observations (t + i) to confirm or reject a change event at t. A change is confirmed in case PChange exceeds a given threshold "chi".
A detailed description is provided in  \href{http://www.mdpi.com/2072-4292/7/5/4973}{Reiche et al. 2015} (Chapter 2.1.4).
}
\author{
Johannes Reiche (Wageningen University)
}
\references{
\href{http://www.mdpi.com/2072-4292/7/5/4973}{Reiche et al. (2015): A Bayesian Approach to Combine Landsat and ALOS PALSAR Time Series for Near Real-Time Deforestation Detection. Remote Sensing. 7(5), 4973-4996; doi:10.3390/rs70504973}
}

