% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getT_change_confirmed.R
\name{getT_change_confirmed}
\alias{getT_change_confirmed}
\title{Get time at which change is confirmed}
\usage{
getT_change_confirmed(bayts)
}
\arguments{
\item{bayts}{"bayts" time series data frame}
}
\value{
time at which change is confirmed
}
\description{
Get time at which change is confirmed from "bayts" time series data frame created with \code{\link{createBayts}} and \code{\link{detectBayts}}.
}
\author{
Johannes Reiche (Wageningen University)
}

