% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getT_change_flagged.R
\name{getT_change_flagged}
\alias{getT_change_flagged}
\title{Get time at which confirmed change was flagged}
\usage{
getT_change_flagged(bayts)
}
\arguments{
\item{bayts}{"bayts" time series data frame}
}
\value{
time at which confirmed change was flagged
}
\description{
Get time at which confirmed change was flagged from "bayts" time series data frame created with \code{\link{createBayts}} and \code{\link{detectBayts}}.
}
\author{
Johannes Reiche (Wageningen University)
}

