% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getT_flagged.R
\name{getT_flagged}
\alias{getT_flagged}
\title{Get time at which unconfirmed change is flagged}
\usage{
getT_flagged(bayts)
}
\arguments{
\item{bayts}{"bayts" time series data frame}
}
\value{
Time at which unconfirmed change is flagged
}
\description{
Get time at which unconfirmed change is flagged from "bayts" time series data frame created with \code{\link{createBayts}} and \code{\link{detectBayts}}.
}
\author{
Johannes Reiche (Wageningen University)
}

