% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBayts.R
\name{plotBayts}
\alias{plotBayts}
\title{Plot original time series including indicated and detected changes (2 time series)}
\usage{
plotBayts(bayts, labL = list(), ylimL = list(), colL = list(),
  xlim = NULL, start = NULL, plotflag = TRUE, plotflag_old = TRUE)
}
\arguments{
\item{bayts}{bayts time series data frame}

\item{labL}{list of y-axis lables}

\item{ylimL}{list of the y-axis limits (x1, x2) for time series 2. The default value is NULL, indicating that the ylim are set to the min and max value of time series 2}

\item{colL}{list of colour for time series}

\item{xlim}{range of x-axis value; default is the range of the time series}

\item{start}{abline for e.g. showing the start of the monitoring period}

\item{plotflag}{plot flagged change}

\item{plotflag_old}{plot old flagged change}
}
\description{
Plot original time series including indicated and detected changes. T.flagged and T.confirmed are plotted
}
\author{
Johannes Reiche (Wageningen University)
}

