% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBaytsPNF.R
\name{plotBaytsPNF}
\alias{plotBaytsPNF}
\title{Plot non-forest probabillty (PNF) time series and indicate changes (from "bayts" time series frame)}
\usage{
plotBaytsPNF(bayts, lab = "PNF", ylim = NULL, col = "black",
  start = NULL)
}
\arguments{
\item{bayts}{bayts time series data frame}

\item{lab}{y-axis lables}

\item{ylim}{y-axis limits. The default value is NULL, indicating that the ylim are set to the min and max value of the ts}

\item{col}{ts colour}

\item{start}{abline for e.g. showing the start of the monitoring period}
}
\description{
Plot non-forest probabillty (PNF) time series and indicate changes (from "bayts" time series object)
T.flagged and T.confirmed are plotted
}
\author{
Johannes Reiche (Wageningen University)
}

