% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotts.R
\name{plotts}
\alias{plotts}
\title{Plot multiple time series with different data ranges}
\usage{
plotts(tsL = list(), ylimL = list(), labL = list(), colL = list(),
  xlim = NULL, pointsL = NULL, title = "")
}
\arguments{
\item{tsL}{list of object(s) of class \code{\link{ts}}}

\item{ylimL}{list of objects(s) of y limits (x1, x2) for tsL. The default value is NULL, indicating that the ylim are set to the min and max value of ts1}

\item{labL}{list of objects(s) of title for tsL}

\item{colL}{list of objects(s) of colour for tsL}

\item{xlim}{range of x-axis value; default is the range of the time series}

\item{pointsL}{list of TRUE and FLASE. (default = NULL = points are plotted for all time series)}

\item{title}{plot title}
}
\description{
Plot multiple time series with different data ranges.
}
\examples{
# load example ts data (Sentinel-1 and Landsat NDVI time series; 09/2014 - 05/2016; Deforestation event in early 2016)
data(s1vv_lndvi_pixel)

# create ts using bfastts (bfast package)
tsS <- bfastts(s1vv_obs,as.Date(s1vv_date),type=c("irregular"))
tsL <- bfastts(lndvi_obs,as.Date(lndvi_date),type=c("irregular"))

# plot ts
plotts(tsL=list(tsL,tsS),labL=list("Landsat NDVI","Sentinel-1 VV [dB]"))
plotts(tsL=list(tsL,tsS),labL=list("Landsat NDVI","Sentinel-1 VV [dB]"),ylimL=list(c(0,1),c(-13,-6)))
}
\author{
Johannes Reiche (Wageningen University)
}

