% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_data.R
\name{s1_landsat_raster.RData}
\alias{s1_landsat_raster.RData}
\title{Sentinel-1 VV and Landsat NDVI raster brick}
\usage{
data(s1_landsat_raster.RData)
}
\description{
Sentinel-1 VV and Landsat NDVI time series raster brick; 09/2014 - 05/2016; Deforestation event in early 2016.  Two univariate time series objects of the class \link{ts}.
}
\examples{
## load test data
load("s1_landsat_raster.RData")

plot(s1,1)
}
\author{
Johannes Reiche
}

