% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_data.R
\name{s1vv_lndvi_pixel.rda}
\alias{s1vv_lndvi_pixel.rda}
\title{Sentinel-1 VV and Landsat NDVI time series example}
\usage{
data(s1vv_lndvi_pixel)
}
\description{
Sentinel-1 VV and Landsat NDVI time series; 09/2014 - 05/2016; Deforestation event in early 2016.  Two univariate time series objects of the class \link{ts}.
}
\examples{
# load test data
data(s1vv_lndvi_pixel)

# create ts using bfastts (bfast package)
tsS <- bfastts(s1vv_obs,as.Date(s1vv_date),type=c("irregular"))
tsL <- bfastts(lndvi_obs,as.Date(lndvi_date),type=c("irregular"))

# plot ts
plotts(tsL=list(tsL,tsS),labL=list("Landsat NDVI","Sentinel-1 VV [dB]"))
plotts(tsL=list(tsL,tsS),labL=list("Landsat NDVI","Sentinel-1 VV [dB]"),ylimL=list(c(0,1),c(-13,-6)))
}
\author{
Johannes Reiche
}

