% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time2date.R
\name{time2date}
\alias{time2date}
\title{Convert time to date}
\usage{
time2date(time, format = "\%Y-\%m-\%d")
}
\arguments{
\item{time}{time}

\item{format}{date format; default="%Y-%m-%d". DOY = "%Y.%j"}
}
\value{
single value, vector or time series object of class "ts"
}
\description{
Convert time to date
}
\author{
Johannes Reiche (Wageningen University)
}
\references{
\href{http://www.mdpi.com/2072-4292/7/5/4973}{Reiche et al. (2015): A Bayesian Approach to Combine Landsat and ALOS PALSAR Time Series for Near Real-Time Deforestation Detection. Remote Sensing. 7(5), 4973-4996; doi:10.3390/rs70504973}
}

