% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPNF.R
\name{calcPNF}
\alias{calcPNF}
\title{Calculating conditional non-forest probabilities}
\usage{
calcPNF(ts, pdf, bwf = c(0, 1))
}
\arguments{
\item{ts}{single value, vector or time series object of class "ts"}

\item{pdf}{"pdf" object describing F and NF distributions.
"pdf" object: pdf\link{1} = pdf type F, pdf\link{2} = pdf type NF, pdf\link{3} and pdf\link{4} = pdf parameter describing F, pdf\link{5} and pdf\link{6} = pdf parameter describing NF. pdf types supported: Gaussian or Weibull.
See Example for details on how to create a "pdf" object.}

\item{bwf}{block weighting function to truncate the NF probability (default = c(0,1) = no truncation)}
}
\value{
single value, vector or time series object of class "ts"
}
\description{
Calculating conditional non-forest probability (PNF). Probabilities are calculated based on pdfs for F and NF (derived from F and NF distributions). (Guassian and Weibull pdf types are supported)
}
\examples{
# NDVI time series
ndvi <- c(0.5,0.7)

# create pdf (Probability density functions) object for forest and non-forest 
pdfType <- c("gaussian","gaussian")
pdfF <-  c(0.85, 0.1) #mean and sd
pdfNF <- c(0.3, 0.2)  #mean and sd
pdf <- c(pdfType,pdfF,pdfNF)

# calculate conditional non-forest probabilities
calcPNF(ndvi,pdf)
}
\references{
\href{http://www.mdpi.com/2072-4292/7/5/4973}{Reiche et al. (2015): A Bayesian Approach to Combine Landsat and ALOS PALSAR Time Series for Near Real-Time Deforestation Detection. Remote Sensing. 7(5), 4973-4996; doi:10.3390/rs70504973}
}
\author{
Johannes Reiche (Wageningen University)
}
